Bot checks which guilds are registered before starting up
